chore: Update file upload logic and handle PDF deletion

+++ src/context/index.jsx
@@ -5,41 +5,68 @@ export const Context = createContext();
 
 const ContextProvider = ({ children }) => {
 	const [input, setInput] = useState("");
-
 	const [loading, setLoading] = useState(false);
-
-	const [response, setResponse] = useState("");
-
-	const [newPrompt, setNewPrompt] = useState("");
-
+	const [response, setResponse] = useState(null);
 	const [pdfMetaData, setPdfMetaData] = useState({
 		name: null,
 		pdf_id: null,
 	});
-
 	const [history, setHistory] = useState([]);
-
+	const [newPrompt, setNewPrompt] = useState("");
+	const [showResult, setShowResult] = useState(false);
 	useEffect(() => {
 		const pdfMetadata = localStorage.getItem("pdfMetaData");
 		if (pdfMetadata) {
 			setPdfMetaData(JSON.parse(pdfMetadata));
 		}
-
-		const chatHistory = localStorage.getItem("chatHistory");
-		if (chatHistory) {
-			setHistory(JSON.parse(chatHistory));
-		}
 	}, []);
+
+	useEffect(() => {
+		const fetchHistory = async (pdf_id) => {
+			try {
+				const response = await axios.get(
+					`${import.meta.env.VITE_BACKEND_URI}/get-history/${pdf_id}`
+				);
+				setHistory(response.data.history);
+			} catch (error) {
+				console.error("Error fetching history:", error);
+			}
+		};
+
+		if (pdfMetaData.pdf_id) {
+			fetchHistory(pdfMetaData.pdf_id);
+		}
+	}, [pdfMetaData.pdf_id]);
 
 	const uploadPdf = async (file) => {
 		setLoading(true);
 		try {
 			const formData = new FormData();
 			formData.append("file", file);
-			if (pdfMetaData.pdf_id !== null && pdfMetaData.pdf_id !== undefined) {
-				await axios.delete(
-					`${import.meta.env.VITE_BACKEND_URI}/delete-pdf/${pdfMetaData.pdf_id}`
-				);
+			// if pdf file is more than 5MB then return
+			if (file.size > 5 * 1024 * 1024) {
+				throw new Error("File size should be less than 5MB");
+			}
+			if (
+				pdfMetaData.pdf_id !== null &&
+				pdfMetaData.pdf_id !== undefined &&
+				pdfMetaData.name !== file.name
+			) {
+				try {
+					await axios.delete(
+						`${import.meta.env.VITE_BACKEND_URI}/delete-pdf/${
+							pdfMetaData.pdf_id
+						}`
+					);
+					console.log("Previous PDF deleted");
+					setPdfMetaData({ name: null, pdf_id: null });
+					localStorage.removeItem("pdfMetaData");
+					setHistory([]);
+					setInput("");
+					setResponse(null);
+				} catch (error) {
+					throw new Error("Error deleting previous PDF");
+				}
 			}
 
 			const res = await axios.post(
@@ -56,7 +83,6 @@ const ContextProvider = ({ children }) => {
 				JSON.stringify({ name: filename, pdf_id })
 			);
 			setHistory([]);
-			localStorage.removeItem("chatHistory");
 		} catch (error) {
 			console.error("Error uploading PDF:", error);
 		} finally {
@@ -66,11 +92,13 @@ const ContextProvider = ({ children }) => {
 
 	const askQuestion = async (question) => {
 		if (loading) return;
-
+		setNewPrompt(question);
+		setResponse(null);
 		setLoading(true);
-		setResponse("");
+		setShowResult(true);
+		setInput("");
 
-		const historySubset = history.slice(-2);
+		const historySubset = history.slice(-3);
 
 		try {
 			const res = await axios.post(
@@ -81,32 +109,35 @@ const ContextProvider = ({ children }) => {
 					history: historySubset,
 				}
 			);
-
+			displayResponseWithAnimation(res.data.answer);
 			const newEntry = {
-				question,
+				question: newPrompt,
 				pdfName: pdfMetaData.name,
 				response: res.data.answer,
 			};
 			const newHistory = [...history, newEntry];
+			setShowResult(false);
+			setResponse(null);
 			setHistory(newHistory);
-			localStorage.setItem("chatHistory", JSON.stringify(newHistory));
-
-			const words = res.data.answer.split(" ");
-			words.forEach((word, index) => delayyPara(index, word + " "));
 		} catch (error) {
 			console.error("Error asking question:", error);
-		} finally {
-			setLoading(false);
-			setNewPrompt("");
-			setResponse("");
 		}
 	};
 
-	//Logic for typing effect : Hoisting
-	const delayyPara = (index, nextWord) => {
-		setTimeout(function () {
+	const delayPara = (index, nextWord) => {
+		setTimeout(() => {
 			setResponse((prev) => prev + nextWord);
-		}, 70 * index);
+		}, 100 * index);
+	};
+
+	const displayResponseWithAnimation = (formattedResponse) => {
+		setLoading(false);
+		setResponse("");
+		const newResponseArray = formattedResponse.split(" ");
+		newResponseArray.forEach((word, index) => {
+			delayPara(index, word + " ");
+		});
+		setInput("");
 	};
 
 	const contextValues = {
@@ -120,10 +151,11 @@ const ContextProvider = ({ children }) => {
 		setResponse,
 		history,
 		setHistory,
-		newPrompt,
-		setNewPrompt,
 		uploadPdf,
 		askQuestion,
+		setNewPrompt,
+		newPrompt,
+		showResult,
 	};
 
 	return <Context.Provider value={contextValues}>{children}</Context.Provider>;
